+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add navigation menu
 - Fix issue with OCSP validation and HTTP 1.1 (StartCOM)
+- Relicense under Affero GPL
 
 ## 1.8